requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
zeep==4.3.1
urllib3==2.2.2
streamlit==1.36.0
